requirements.txt
telethon
python-aliexpress-api
python-dotenv